+++ wazo_confd/plugins/endpoint_sip/resource.py
@@ -1,4 +1,4 @@
-# Copyright 2015-2020 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2022 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 from flask import url_for, request
@@ -35,7 +35,10 @@ class _BaseSipList(ListResource):
         return super().get()
 
     def post(self):
-        form = self.schema().load(request.get_json())
+        return super().post()
+
+    def _post(self, body):
+        form = self.schema().load(body)
         form = self.add_tenant_to_form(form)
 
         templates = []